+++ fabric/src/main/java/dev/greenhouseteam/enchantmentdisabletag/mixin/fabric/ReloadableServerResourcesMixin.java
@@ -1,0 +1,17 @@
+package dev.greenhouseteam.enchantmentdisabletag.mixin.fabric;
+
+import dev.greenhouseteam.enchantmentdisabletag.EnchantmentDisableTag;
+import net.minecraft.core.RegistryAccess;
+import net.minecraft.server.ReloadableServerResources;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+@Mixin(ReloadableServerResources.class)
+public class ReloadableServerResourcesMixin {
+    @Inject(method = "updateRegistryTags(Lnet/minecraft/core/RegistryAccess;)V", at = @At("TAIL"))
+    private void enchantmentdisabletag$setToReloadCreativeTabsServer(RegistryAccess registryAccess, CallbackInfo ci) {
+        EnchantmentDisableTag.setReloaded();
+    }
+}