+++ fabric/src/main/java/dev/greenhouseteam/enchantmentdisabletag/mixin/fabric/client/ClientPacketListenerMixin.java
@@ -9,9 +9,9 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Mixin(ClientPacketListener.class)
-public class ClientPacketListenerMixin {
+public abstract class ClientPacketListenerMixin {
     @Inject(method = "handleUpdateTags", at = @At("TAIL"))
-    private void enchantmentdisabletag$updateTagValue(ClientboundUpdateTagsPacket packet, CallbackInfo ci) {
+    private void enchantmentdisabletag$setToReloadCreativeTabsClient(ClientboundUpdateTagsPacket packet, CallbackInfo ci) {
         EnchantmentDisableTag.setReloaded();
     }
 }