+++ CHANGELOG.md
@@ -1,1 +1,1 @@
-- Initial 1.20.1 release.
+- Allowed creative mode tab code to reload on the server.